pyproject.toml
[project]
name = "tv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "async-timeout>=5.0.1",
    "google-genai>=0.5.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "ruff>=0.14.10",
    "samsungtvws[async,cli,encrypted]>=3.0.0",
]
